% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{aggregate_polygons}
\alias{aggregate_polygons}
\title{Aggregate the attributes of the polygons grouped by some criteria.}
\usage{
aggregate_polygons(spdf, gridSpdf, by = NULL, minArea = 0.5, colNames,
  colFun)
}
\arguments{
\item{spdf}{A `\code{\link[sp]{SpatialPolygonsDataFrame}}` object with the
polygons to be cropped.}

\item{gridSpdf}{A `\code{\link[sp]{SpatialPolygonsDataFrame}}` object with the
grid as created by \code{\link{make_grid}}.}

\item{by}{Character vector with the name of the columns used to group polygons.}

\item{minArea}{A numeric between 0 and 1 indicating the minimum proportion of area overlay between polygons in `spdf` and a pixel. Grid pixels whose area do not cover the minimum in proportion will be dropped.}

\item{colNames}{Character vector with the name of the columns that will be aggregated.}

\item{colFun}{Function vector with the function that will be applied to each column.}
}
\description{
Aggregate the attributes of the polygons grouped by some criteria.
}
