% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_ys}
\alias{apply_ys}
\title{Apply a function for each combination of factors.}
\usage{
apply_ys(X, FUN, by = c("year", "site"))
}
\arguments{
\item{X}{A data.frame with named columns including at least those in `by`.}

\item{FUN}{A function expecting a data.frame and returning an atom.}

\item{by}{A vector of character strings with the names of the columns used to split the dataset.}
}
\value{
A vector with length equal to the number of rows in the data.frame containing the values returned by `FUN`.
}
\description{
Apply a function for each combination of factors.
}
\examples{
apply_ys(yieldExtra, function(df) {
  bearing(df$x, df$y)
})
apply_ys(yieldExtra, function(df) {
  hav_distance_km(df$x, df$y)
})
}
