% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{crop_polygon}
\alias{crop_polygon}
\title{Crop out one polygone from another.}
\usage{
crop_polygon(spFrom, spTo)
}
\arguments{
\item{spFrom}{A \code{\link[sp]{Polygon}} from which `polyTo` will be
removed.}

\item{spTo}{A \code{\link[sp]{Polygon}} that will be cropped out of `spFrom`.}
}
\value{
A \code{\link[sp]{Polygon}} object with the region of `spFrom` that
  is not within `spTo`.
}
\description{
Crop out one polygone from another.
}
