% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{euc_distance_m}
\alias{euc_distance_m}
\title{Compute the distance between two coordinates in meters}
\usage{
euc_distance_m(x, y)
}
\arguments{
\item{x}{A numeric vector with easting position in meters.}

\item{y}{A numeric vector with northing position in meters.}
}
\value{
A numeric vector with the distance in meters
}
\description{
Compute the distance between two coordinates in meters
}
