% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_polygon_area_m2}
\alias{get_polygon_area_m2}
\title{Get the area of child polygons.}
\usage{
get_polygon_area_m2(spdf)
}
\arguments{
\item{sp}{A `\code{\link[sp]{Polygons}}` object.}
}
\value{
A numeric vector with the area in square meters of all the polygons
in the input object.
}
\description{
Get the area of child polygons.
}
