% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_polygon_id}
\alias{get_polygon_id}
\title{Get ID of child polygons.}
\usage{
get_polygon_id(sp)
}
\arguments{
\item{sp}{A `\code{\link[sp]{Polygons}}` object.}
}
\value{
A character vector with the ID of all the polygons conforming
the input object.
}
\description{
Get ID of child polygons.
}
