% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{grain_market_moisture}
\alias{grain_market_moisture}
\title{Return the nominal market moisture content of a crop.}
\usage{
grain_market_moisture(cropString)
}
\arguments{
\item{cropString}{A character string, or a vector of character strings, with
the name of the crop. Use either "Corn" or "Soybeans".}
}
\value{
The nominal market moisture content (% hundreds).
}
\description{
Return the nominal market moisture content of a crop.
}
\examples{
grain_market_moisture(c("Corn", "Soybeans"))
}
