% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{grain_test_mass}
\alias{grain_test_mass}
\title{Return the test mass (weight) of a crop.}
\usage{
grain_test_mass(cropString)
}
\arguments{
\item{cropString}{A character string, or a vector of character strings, with the name of the crop. Use either "Corn" or "Soybeans".}
}
\value{
The test weight of the crop in pounds per bushel.
}
\description{
Return the test mass (weight) of a crop.
}
\examples{
grain_test_mass(c("Corn", "Soybeans"))
}
