% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{hav_distance_m}
\alias{hav_distance_m}
\title{Compute the distance between two coordinates in meters}
\usage{
hav_distance_m(x, y)
}
\arguments{
\item{x}{A numeric vector with longitudes.}

\item{y}{A numeric vector with latitudes.}
}
\value{
A numeric vector with the distance in meters
}
\description{
Compute the distance between two coordinates in meters
}
\author{
Jaap https://stackoverflow.com/a/23095329
}
