% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{lognormal_to_normal}
\alias{lognormal_to_normal}
\title{Transform the parameter value from lognormal to normal distribution.}
\usage{
lognormal_to_normal(mean, var, what)
}
\arguments{
\item{mean}{A numeric with the mean of the lognormal distribution.}

\item{var}{A numeric with the variance of the lognormal distribution.}

\item{what}{A string with the name of the value to return. It can be `mean`,
`var`, or `median`.}
}
\value{
A numeric with the transformed value.
}
\description{
Transform the parameter value from lognormal to normal distribution.
}
