% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{make_bounding_box}
\alias{make_bounding_box}
\title{Find the vertices of the polygon associated with a moving vehicle.}
\usage{
make_bounding_box(x, y, w)
}
\arguments{
\item{x}{A numeric vector with the latitude of the vehicle locations.}

\item{y}{A numeric vector with the longitude of the vehicle locations}

\item{w}{A numeric vector with the width of the vehicle IN METERS. Again, `w`
MUST BE IN METERS.}
}
\value{
A matrix with eight columns with the coordinates of the four
  vertices, where 11 and 12 represent the points closer to the initial
  location and 21 and 22 represent the pointers closer to the final location.
}
\description{
Find the vertices of the polygon associated with a moving vehicle.
}
\note{
`w` MUST BE IN METERS.
}
