% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{make_grid}
\alias{make_grid}
\title{Create a grid covering the given polygons.}
\usage{
make_grid(spdf, n = NULL, width = NULL, height = NULL, minArea = 0,
  regular = TRUE)
}
\arguments{
\item{spdf}{A `\code{\link[sp]{SpatialPolygonsDataFrame}}` object with the
polygons to be covered by the grid.}

\item{n}{The approximate number of polygons that the grid should contain.}

\item{width}{The width of the grid polygons in meters.}

\item{height}{The height of the grid polygons in meters.}

\item{minArea}{A numeric between 0 and 1 indicating the minimum proportion of area overlay between polygons in `spdf` and a pixel. Grid pixels whose area do not cover the minimum in proportion will be dropped.}

\item{regular}{If TRUE (default), the union of all the elements in the grid
will form a regular polygons. If FALSE, grid polygons not intercepting with
any of the input polygons will be dropped. NOTE: this is not about the
elements of of the grid having a regular shape!}
}
\value{
A `\code{\link[sp]{SpatialPolygonsDataFrame}}` object with the
polygons conforming the grid.
}
\description{
Note that `width` and `height` take priority over `n` if all are non-null.
}
\references{
Loosely based on
https://ecosystems.psu.edu/research/labs/walter-lab/manual/chapter1/1.9-creating-a-square-polygon-grid-over-a-study-area
}
