% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{make_grid_by_size}
\alias{make_grid_by_size}
\title{Title}
\usage{
make_grid_by_size(spdf, width, height)
}
\arguments{
\item{spdf}{A `\code{\link[sp]{SpatialPolygonsDataFrame}}` with the polygons
that will be covered by the grid.}

\item{width}{Grid width in meters.}

\item{height}{Grid height in meters.}
}
\value{
A data.frame with two columns.
}
\description{
Title
}
