% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{make_vehicle_polygons}
\alias{make_vehicle_polygons}
\title{Create polygons from a data.frame with centroid coordinates.}
\usage{
make_vehicle_polygons(df, proj4string)
}
\arguments{
\item{df}{A data.frame with columns `x`, `y`, and `swath`. The latter
represents the box width and must be in meters.}
}
\value{
A `\code{\link[sp]{SpatialPolygonsDataFrame}}` with one rectangle per
row.
}
\description{
Create a `\code{\link[sp]{SpatialPolygonsDataFrame}}` object from the `sp`
package based on the data.frame given as an input. For each row in the input
data.frame, a rectangle is constructed using the current and next positions
as well as the swath width.
}
\note{
`swath` MUST BE IN METERS.
}
