% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_movement}
\alias{plot_movement}
\title{Plot the vehicle in movement.}
\usage{
plot_movement(sp, endpoints, main = NA, subtitle = NA, cex = 1,
  boxes = TRUE, vertices = TRUE, locations = TRUE, arrows = TRUE,
  desc = TRUE, legend = "topright", ...)
}
\arguments{
\item{sp}{A `spatialPolygons` object containing the vehicle polygons.}

\item{endpoints}{A data.frame containing the following columns:
x0, x1, y0, y1}

\item{main}{A character string with the plot title (optional).}

\item{subtitle}{A character string with the plot subtitle (optional).}

\item{cex}{The cex param.}

\item{boxes}{If TRUE, a rectangle representing the vehicle are drawn.}

\item{vertices}{If TRUE, points representing the vertices of the vehicle are drawn.}

\item{locations}{If TRUE, points representing the initial and final locations are drawn.}

\item{arrows}{If TRUE, arrows pointing from the initial to the final locations are drawn.}

\item{desc}{If TRUE, the bearing degrees and the distance in meters are printed.}

\item{legend}{If not NULL, the position of the legend.}
}
\value{
Nothing.
}
\description{
Plot the vehicle in movement.
}
