% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{reshape_polygons}
\alias{reshape_polygons}
\title{Reshape overlapping polygons.}
\usage{
reshape_polygons(spdf, verbose = TRUE)
}
\arguments{
\item{spdf}{A `\code{\link[sp]{SpatialPolygonsDataFrame}}` returned by
`\code{\link{make_vehicle_polygons}}`.}

\item{verbose}{If TRUE, progress will be printed.}
}
\value{
The input object with reshaped polygons.
}
\description{
Reshape overlapping polygons.
}
