% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ritas-package}
\alias{ritas}
\alias{ritas-package}
\title{ritas: RITAS: Smooth maps for overlapping, unequally-shaped, irregularly-sized, time-ordered, areal spatial units}
\description{
An implementation of .
}
\details{
See the Introduction vignette: vignette("introduction", package = "yieldMaps")
Additionally, you may start with the manual help for ?ritas.
}
\author{
\strong{Maintainer}: Luis Damiano \email{ldamiano@iastate.edu} (\href{https://orcid.org/0000-0001-9107-0706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jarad Niemi \email{niemi@iastate.edu} (\href{https://orcid.org/0000-0002-5079-158X}{ORCID})
}

}
\keyword{internal}
