% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{yield_equation}
\alias{yield_equation}
\title{Compute the yield in bushels per acre (bu/ac)}
\usage{
yield_equation(massFlowRate, time, distance, swathWidth, testMass,
  harvestMoisture, marketMoisture = 0)
}
\arguments{
\item{massFlowRate}{mass flow rate (lb/sec)}

\item{time}{logging interval (sec)}

\item{distance}{distance traveled between logged data points (ft)}

\item{swathWidth}{header cut width (ft)}

\item{testMass}{grain density (lb/bu)}

\item{harvestMoisture}{moisture content (percentage)}

\item{marketMoisture}{market moisture content (percentage). Set to zero if you want the dry yield.}
}
\value{
Yield (bu/ac)
}
\description{
Compute the yield in bushels per acre (bu/ac)
}
\references{
Best Management Practices for Collecting AccurateYield Data and Avoiding Errors During Harvest http://extensionpublications.unl.edu/assets/pdf/ec2004.pdf
}
