% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{yield_equation_mgha}
\alias{yield_equation_mgha}
\title{Compute the yield in megagrams per hectare (mg/ha)}
\usage{
yield_equation_mgha(massKg, areaM2)
}
\arguments{
\item{massKg}{mass (kg)}

\item{areaM2}{area (square meters)}
}
\value{
Yield (mg/ha)
}
\description{
Compute the yield in megagrams per hectare (mg/ha)
}
