% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{yieldMaps-package}
\alias{yieldMaps}
\alias{yieldMaps-package}
\title{yieldMaps: Title}
\description{
Description.
}
\details{
See the Introduction vignette: vignette("introduction", package = "yieldMaps")
Additionally, you may start with the manual help for ?ritas.
}
\author{
\strong{Maintainer}: Luis Damiano \email{ldamiano@iastate.edu} (0000-0001-9107-0706) [copyright holder]

Authors:
\itemize{
  \item Jarad Niemi \email{niemi@iastate.edu}
}

}
\keyword{internal}
